+++ t-tabular/CHANGELOG.md
@@ -5,6 +5,7 @@ v2.3.0
 ---
 * Space trailing is now accessible for all input types.
 * Dialog layout update, to provide more space for check boxes.
+* For XSL: Option to use advanced value detection for better handling of integers and dates.
 
 v2.2.2
 ---